% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{gfilmm-imports}
\alias{gfilmm-imports}
\alias{ewcdf}
\alias{quantile.ewcdf}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages.
  Follow the links to their documentation:
  \code{\link[spatstat:ewcdf]{ewcdf}},
  \code{\link[spatstat:quantile.ewcdf]{quantile.ewcdf}}.
}
