% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfilmm.R
\name{gfilmm}
\alias{gfilmm}
\title{Generalized fiducial inference}
\usage{
gfilmm(yl, yu, FE, RE, N, thresh = N/2)
}
\arguments{
\item{yl, yu}{data}

\item{FE}{fixed effects}

\item{RE}{random effects}

\item{N}{number of simulations}

\item{thresh}{threshold, default \code{N/2}}
}
\value{
A list with two components: \code{VERTEX} and \code{WEIGHT}.
}
\description{
Samples the fiducial distributions.
}
\examples{
h <- 0.01
yl <- npk$yield-h
yu <- npk$yield+h
FE <- as.matrix(rep(1,nrow(npk))) # intercept
RE <- data.frame(block = npk$block)
gfi <- gfilmm(yl, yu, FE, RE, N=5000)
library(spatstat) # to use ewcdf 
f <- ewcdf(gfi$VERTEX[1,], gfi$WEIGHT)
plot(f, xlim = c(40,65))
quantile(f, c(0.025,0.975))
}
\references{
J. Cisewski and J.Hannig. 
\emph{Generalized fiducial inference for normal linear mixed models}. 
The Annals of Statistics 2012, Vol. 40, No. 4, 2102–2127.
}
