% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{gfiSummary}
\alias{gfiSummary}
\title{Summary of fiducial distributions}
\usage{
gfiSummary(gfi, conf = 0.95)
}
\arguments{
\item{gfi}{output of \code{\link{gfilmm}}}

\item{conf}{confidence level}
}
\value{
Summary statistics in a matrix.
}
\description{
Summary of fiducial distributions
}
\examples{
data(KM41)
h <- 0.005
gfi <- gfilmm(~ cbind(y-h, y+h), ~ 1, ~ Batch, N=500)
gfiSummary(gfi)
}
