% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{KM41}
\alias{KM41}
\title{Krishnamoorthy & Mathew's example 4.1}
\format{
A data frame with 25 rows and 2 columns.
}
\usage{
data(KM41)
}
\description{
The dataset used in Krishnammorthy & Mathew's example 4.1.
}
\examples{
data(KM41)
str(KM41)
table(KM41$Batch)
}
\references{
Krishnamoorthy and Mathew, Statistical Tolerance Regions, Wiley 2009.
}
\keyword{data}
