% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pHdata}
\alias{pHdata}
\title{pH dataset}
\format{
A data frame with 160 rows and 4 columns. 
Column \code{SIRE} is a factor nested in column \code{DAM}.
}
\usage{
data(pHdata)
}
\description{
A dataset from ?? (I don't remember).
}
\examples{
data(pHdata)
str(pHdata)
table(droplevels(pHdata[pHdata$DAM=="D1","SIRE"]))
table(droplevels(pHdata[pHdata$DAM=="D2","SIRE"]))
table(droplevels(pHdata[pHdata$DAM=="D3","SIRE"]))
}
\keyword{data}
