% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfilmm.R
\name{gfilmm}
\alias{gfilmm}
\title{Generalized fiducial inference}
\usage{
gfilmm(y, fixed, random, data, N, thresh = N/2)
}
\arguments{
\item{y}{a right-sided formula of the form \code{~ cbind(lower,upper)} for 
the interval data}

\item{fixed}{a right-sided formula for the fixed effects}

\item{random}{a right-sided formula for the random effects, or \code{NULL} 
for no random effect}

\item{data}{the data, a dataframe}

\item{N}{desired number of simulations}

\item{thresh}{threshold, default \code{N/2}; for experts only}
}
\value{
A list with three components: \code{VERTEX}, \code{WEIGHT} and 
  \code{ESS}.
}
\description{
Samples the fiducial distributions.
}
\examples{
h <- 0.01
gfi <- gfilmm(~ cbind(yield-h, yield+h), ~ 1, ~ block, data = npk, N=5000)
# fiducial cumulative distribution function of the intercept:
library(spatstat) # to use ewcdf 
f <- ewcdf(gfi$VERTEX["(Intercept)",], gfi$WEIGHT)
plot(f, xlim = c(40,65))
# fiducial confidence interval of the intercept:
quantile(f, c(0.025,0.975))
# fiducial density function of the intercept:
library(kde1d)
kfit <- kde1d(gfi$VERTEX["(Intercept)",], weights = gfi$WEIGHT)
curve(dkde1d(x, kfit), from = 45, to = 65)
}
\references{
J. Cisewski and J.Hannig. 
\emph{Generalized fiducial inference for normal linear mixed models}. 
The Annals of Statistics 2012, Vol. 40, No. 4, 2102–2127.
}
