% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfilmm.R
\name{gfilmm}
\alias{gfilmm}
\title{Generalized fiducial inference}
\usage{
gfilmm(y, fixed, random, data, N, thresh = N/2)
}
\arguments{
\item{y}{a formula of the form \code{~ cbind(lower,upper)} for the interval 
data}

\item{fixed}{formula for the fixed effects}

\item{random}{formula for the random effects}

\item{data}{XXX}

\item{N}{number of simulations}

\item{thresh}{threshold, default \code{N/2}; for experts only}
}
\value{
A list with two components: \code{VERTEX} and \code{WEIGHT}.
}
\description{
Samples the fiducial distributions.
}
\examples{
h <- 0.01
# gfi <- gfilmm(~ cbind(yield-h, yield+h), ~ 1, ~ block, data = npk, N=500)
# library(spatstat) # to use ewcdf 
# f <- ewcdf(gfi$VERTEX[1,], gfi$WEIGHT)
# plot(f, xlim = c(40,65))
# quantile(f, c(0.025,0.975))

##########################################################################
SimAV2mixed <- function(I, J, Kij, mu=0, alphai, sigmaO=1,
                        sigmaPO=1, sigmaE=1, factor.names=c("Part","Operator"),
                        resp.name="y", keep.intermediate=FALSE){
  Operator <- rep(1:J, each=I)
  Oj <- rep(rnorm(J, 0, sigmaO), each=I)
  Part <- rep(1:I, times=J)
  Pi <- rep(alphai, times=J)
  POij <- rnorm(I*J, 0, sigmaPO)
  simdata0 <- data.frame(Part, Operator, Pi, Oj, POij)
  simdata0$Operator <- factor(simdata0$Operator)
  levels(simdata0$Operator) <- 
    sprintf(paste0("\%0", floor(log10(J))+1, "d"), 1:J)
  simdata0$Part <- factor(simdata0$Part)
  levels(simdata0$Part) <- sprintf(paste0("\%0", floor(log10(I))+1, "d"), 1:I)
  simdata <- 
    as.data.frame(
      sapply(simdata0, function(v) rep(v, times=Kij), simplify=FALSE))
  Eijk <- rnorm(sum(Kij), 0, sigmaE)
  simdata <- cbind(simdata, Eijk)
  simdata[[resp.name]] <- mu + with(simdata, Oj+Pi+POij+Eijk)
  levels(simdata[,1]) <- paste0("A", levels(simdata[,1]))
  levels(simdata[,2]) <- paste0("B", levels(simdata[,2]))
  names(simdata)[1:2] <- factor.names
  if(!keep.intermediate) simdata <- simdata[,c(factor.names,resp.name)]
  simdata
}


set.seed(666)  
I = 2; J = 3; Kij = rep(5,I*J)
alphai <- c(3, -3)
sigmaO <- 2
sigmaPO <- 1
sigmaE <- sqrt(3)
dat <- SimAV2mixed(I, J, Kij, mu = 0, alphai = alphai, 
                   sigmaO = sigmaO, sigmaPO = sigmaPO, sigmaE = sigmaE)


library(gfilmm)
h <- 0.01
gfi <- gfilmm(~ cbind(y-h,y+h), ~ Part, ~ Operator, data = dat, 
              N = 100, thresh=0)
}
\references{
J. Cisewski and J.Hannig. 
\emph{Generalized fiducial inference for normal linear mixed models}. 
The Annals of Statistics 2012, Vol. 40, No. 4, 2102–2127.
}
