% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfilmm.R
\name{gfilmm}
\alias{gfilmm}
\title{Generalized fiducial inference}
\usage{
gfilmm(y, fixed, random, data, N, thresh = N/2)
}
\arguments{
\item{y}{a formula of the form \code{~ cbind(lower,upper)} for the interval 
data}

\item{fixed}{formula for the fixed effects}

\item{random}{formula for the random effects}

\item{N}{number of simulations}

\item{thresh}{threshold, default \code{N/2}; for experts only}
}
\value{
A list with two components: \code{VERTEX} and \code{WEIGHT}.
}
\description{
Samples the fiducial distributions.
}
\examples{
h <- 0.01
gfi <- gfilmm(~ cbind(yield-h, yield+h), ~ 1, ~ block, data = npk, N=500)
library(spatstat) # to use ewcdf 
f <- ewcdf(gfi$VERTEX[1,], gfi$WEIGHT)
plot(f, xlim = c(40,65))
quantile(f, c(0.025,0.975))
}
\references{
J. Cisewski and J.Hannig. 
\emph{Generalized fiducial inference for normal linear mixed models}. 
The Annals of Statistics 2012, Vol. 40, No. 4, 2102–2127.
}
