% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{gfiPredictive}
\alias{gfiPredictive}
\title{Generalized fiducial predictive distributions}
\usage{
gfiPredictive(gfi, newdata)
}
\arguments{
\item{gfi}{a \code{\link{gfilmm}} object}

\item{newdata}{dataframe in which to look for variables with which to predict}
}
\value{
The simulations in a dataframe.
}
\description{
Simulations of the generalized fiducial predictive 
  distributions.
}
\examples{
gfi <- gfilmm(~ cbind(yield-0.1, yield+0.1), ~ N, ~ P, npk, 20)
gfiPredictive(gfi, data.frame(N = "0", P = "0"))
}
